third_party/WebKit/LayoutTests/http/tests/inspector/network/load-resource-when-paused.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../debugger-test.js"></script>
<script>

function throwAnException()
{
    return unknown_var;
}

function handleClick()
{
    throwAnException();
}

var test = function()
{
    InspectorTest.startDebuggerTest(step1);

    function step1()
    {
        DebuggerAgent.setPauseOnExceptions(WebInspector.DebuggerModel.PauseOnExceptionsState.PauseOnUncaughtExceptions);
        InspectorTest.evaluateInPage("setTimeout(handleClick, 0)");
        InspectorTest.waitUntilPaused(step2);
    }

    function step2()
    {
        var mainResourceURL = WebInspector.resourceTreeModel.mainFrame.url;
        var resourceURL = WebInspector.ParsedURL.completeURL(mainResourceURL, "resources/resource.php");
        NetworkAgent.loadResourceForFrontend(WebInspector.resourceTreeModel.mainFrame.id, resourceURL, undefined, resourceContentLoaded);
    }

    function resourceContentLoaded(error, statusCode, headers, content)
    {
        InspectorTest.assertTrue(!error, "There shouldn't have been an error on resource load: " + error);
        InspectorTest.addResult("Resource content loaded: " + content);
        InspectorTest.completeDebuggerTest();
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that loading external resources works when inspected page is paused.
</p>

</body>
</html>
