third_party/WebKit/LayoutTests/http/tests/inspector/network/load-resource-when-paused-expected.txt

CONSOLE ERROR: line 9: Uncaught ReferenceError: unknown_var is not defined
Tests that loading external resources works when inspected page is paused.

Script execution paused.
Resource content loaded: Hello world
Script execution resumed.
